+++ CHANGES.rst
@@ -37,6 +37,7 @@ Bugfixes
 - Fix date picker showing January instead of the selected month when the user
   language is not English (:issue:`7471`, :pr:`7476`, thanks :user:`foxbunny`)
 - Do not show "Clone Abstract" icon outside management area (:pr:`7493`)
+- Allow rejecting editables w/ missing required files (:pr:`7524`)
 
 Accessibility
 ^^^^^^^^^^^^^